+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ## 0.2.0 (2023-11-25)
 - API functions now return Object instead of Value
 - add coprMonitorProject
-- add coprGetBuildList
+- coprGetBuildList replaces coprGetBuildPackageList
 - haddock urls updated to point to main branch
 
 ## 0.1.0 (2021-12-29)